Created /authCheck Route
Created /authCheck which use express-jwt to verify the jwt and console log
the payload.

Added a configuration file expressTest.config.js which contains login
information for a test user in addition the standard expressConfig.  Kept
as seperate files in case future requires different settings.

Updated indextTest to login a user and get a JWT that can be passed to the
other route tests.